pyproject.toml
[project]
name = "myrag-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Antonio Morabito",email = "antoniomorabitositorus@gmail.com"}
]
readme = "README.md"

[tool.poetry]
packages = [{include = "myrag_chatbot", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "*"
langchain-google-genai ="2.1.2,<3.0.0"
cryptography =">=44.0.2,<45.0.0"
langchain-community = "*"
langchain-ollama = "*"
langchain-core = "^0.3"
langchain-openai = "*"
langchain-text-splitters = "*"
chainlit = "^1.2.0"
langchainhub = "^0.1.21"
langchain-chroma = "^0.2.2"
python-dotenv = "*"
sentence-transformers = "^4.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
